:construction: Move commentfield to under comments
